Modularización del código y documentación del mismo
